✨ Refactor: 홈 컴포넌트 분리 및 타입 정리
홈 컴포넌트 분리 및 타입 정리
