+++ src/stores/scanResultStore.ts
@@ -1,42 +1,14 @@
 import { create } from "zustand";
 import apiClient from "@utils/apiClient";
-
-export interface TrashPart {
-  name: string;
-  typeCode: string;
-  typeName: string;
-}
-
-export interface Location {
-  id: string;
-  sido: string;
-  sigungu: string;
-  eupmyeondong: string;
-}
-
-export interface ApiScanResult {
-  id: number;
-  imageUrl: string;
-  name: string;
-  summary: string;
-  itemName: string;
-  typeCode: string;
-  typeName: string;
-  guideSteps: string[];
-  cautionNote: string | null;
-  parts: TrashPart[];
-  createdAt: string;
-  location?: Location;
-  days?: string[];
-}
+import type { ApiTrashDetail } from "@types";
 
 interface ScanResultState {
-  resultsById: { [key: number]: ApiScanResult };
-  currentResult: ApiScanResult | null;
-  setResultFromNavigation: (result: ApiScanResult) => void;
+  resultsById: { [key: number]: ApiTrashDetail };
+  currentResult: ApiTrashDetail | null;
+  setResultFromNavigation: (result: ApiTrashDetail) => void;
   fetchResultById: (id: number) => Promise<void>;
   clearCurrentResult: () => void;
-  updateCurrentResult: (newResult: ApiScanResult) => void;
+  updateCurrentResult: (newResult: ApiTrashDetail) => void;
 }
 
 export const useScanResultStore = create<ScanResultState>((set, get) => ({
@@ -56,7 +28,7 @@ export const useScanResultStore = create<ScanResultState>((set, get) => ({
       set({ currentResult: cachedResult });
     }
     try {
-      const response = await apiClient.get<{ data: ApiScanResult }>(
+      const response = await apiClient.get<{ data: ApiTrashDetail }>(
         `/api/v1/trash/${id}`
       );
       const result = response.data.data;
