+++ src/types/index.ts
@@ -1,4 +1,5 @@
-export * from "./home/schedule";
-export * from "./home/ranking";
-export * from "./home/revision";
-export * from "./location/location";
+export * from "./schedule";
+export * from "./ranking";
+export * from "./revision";
+export * from "./location";
+export * from "./trash";
