+++ src/types/schedule.ts
